fixed countByCriteria methods to Long, added count tests to servicetests

+++ iot-industrial-internet/src/main/java/fi/iot/iiframework/daos/GenericDAO.java
@@ -53,5 +53,5 @@ public interface GenericDAO<T, ID extends Serializable> {
 
     public List<T> findByCriteriaFromTo(int from, int to, List<Criterion> list);
     
-    public int countByCriteria(List<Criterion> list);
+    public Long countByCriteria(List<Criterion> list);
 }
