+++ iot-industrial-internet/src/main/java/fi/iot/iiframework/services/GenericService.java
@@ -23,7 +23,7 @@ public interface GenericService<T, ID extends Serializable> {
 
     public List<T> getAll();
     
-    public int countByCriteria(List<Criterion> criterion);
+    public Long countByCriteria(List<Criterion> criterion);
     
-    public int count();
+    public Long count();
 }
