+++ iot-industrial-internet/src/test/java/fi/iot/iiframework/services/domain/ReadoutServiceTest.java
@@ -110,12 +110,12 @@ public class ReadoutServiceTest {
     
     @Test
     public void readoutsCanBeCounted() {
-        assertEquals(3, service.count());
+        assertEquals(3, (long) service.count());
     }
     
     @Test
     public void readoutsCanBeCountedBySensor() {
-        assertEquals(2, service.countBy(s1));
+        assertEquals(2, (long) service.countBy(s1));
     }
 
 }
